+++ CHANGELOG.md
@@ -8,8 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - [Core] Change `<SearchInput>` behavior:
     - Can be controlled now, via props `value`, `onChange` and `onReset`.
-    - No longer trigger `onSearch` when input blur by default. You can enable this behavior by setting prop `searchWhenInputBlur` be `true`
-    - New prop `searchWhenInputChange`, when it is `true`, `onSearch` will be triggered every time after input changed. The default value is `false`.
+    - No longer trigger `onSearch` when input blur by default. You can enable this behavior by setting prop `searchOnInputBlur` be `true`
+    - New prop `searchOnInputChange`, when it is `true`, `onSearch` will be triggered every time after input changed. The default value is `false`.
     - New prop `blockDuplicateValueSearch`, when it is `true`, `onSearch` will not be triggerd if input value is same with last time searching.
     - New prop `blockEmptyValueSearch`, when it is `true`, `onSearch` will not be triggerd if input value is empty.
     - Rename prop `input` to be `inputProps`.